+++ gui/__init__.py
@@ -1,0 +1,1 @@
+from .core import EmailChatbotGUI, GUI_AVAILABLE, can_initialize_gui
